AP2/requirements.txt
dj-database-url==0.5.0
Django==3.2
gunicorn==20.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2019.3
whitenoise==5.2.0
httplib2==0.14.0
urllib3==1.25.8
requests==2.22.0
django-ninja==0.12.3
djangorestframework==3.12.4